backend/requirements.txt
fastapi
uvicorn
slowapi
httpx
sqlalchemy
pydantic
pydantic-settings
alembic
PyJWT
python-dotenv
psycopg[binary]
pytest
pytest-asyncio
ruff
paho-mqtt
firebase-admin
twilio
python-multipart
